requirements.txt
amqp==5.2.0
asgiref==3.8.1
async-timeout==4.0.3
billiard==4.2.0
celery==5.3.6
certifi==2024.7.4
charset-normalizer==3.3.2
click==8.1.7
click-didyoumean==0.3.1
click-plugins==1.1.1
click-repl==0.3.0
cron-descriptor==1.4.3
Django>=4.2,<5.0
django-celery-beat==2.5.0
django-rest-framework==0.1.0
django-timezone-field==7.0
djangorestframework==3.14.0
gunicorn==22.0.0
idna==3.7
kombu==5.3.7
packaging==24.1
prompt_toolkit==3.0.47
psutil==5.9.7
python-crontab==3.2.0
python-dateutil==2.9.0.post0
python-dotenv==1.0.0
redis==5.0.1
requests==2.31.0
six==1.16.0
sqlparse==0.5.0
typing_extensions==4.12.2
tzdata==2024.1
urllib3==2.2.2
vine==5.1.0
wcwidth==0.2.13
